requirements.txt
# Core Framework
flask==2.3.3
werkzeug==2.3.7

# AI/ML Libraries
tensorflow==2.13.0
opencv-python-headless==4.8.1.78
numpy==1.23.5
scikit-learn==1.3.0

# Data Processing
pandas==2.0.3
pillow==10.0.0

# Visualization
plotly==5.15.0
matplotlib==3.7.2
seaborn==0.12.2

# Web Server (Production)
gunicorn==21.2.0

# Utilities
requests==2.31.0
python-dotenv==1.0.0

# Database
flask-sqlalchemy==3.0.5

# CORS Support
flask-cors==4.0.0
